move WebotsSupervisor functionality to a new class
